Make videos more consistent between granules and collections.
