+++ client/src/detail/VideoView.jsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import Video from '../common/Video'
-import {govExternalYouTubeMsg} from '../utils/urlUtils'
 
 const styleVideoView = {
   display: 'flex',
@@ -37,16 +36,6 @@ const styleListElement = {
 const styleListElementSelected = {
   backgroundColor: '#b0d1ea',
   color: '#222',
-}
-
-const styleDisclaimer = {
-  color: '#f9f9f9',
-  border: '3px solid #3a3a3a',
-  borderCollapse: 'collapse',
-  backgroundColor: '#1a1a1a',
-  margin: 0,
-  padding: '1.618em',
-  fontStyle: 'italic',
 }
 
 export default class VideoView extends React.Component {
@@ -118,7 +107,6 @@ export default class VideoView extends React.Component {
           <ul style={styleList}>{titleList}</ul>
           {embeddedVideos}
         </div>
-        <div style={styleDisclaimer}>Disclaimer: {govExternalYouTubeMsg}</div>
       </div>
     )
   }
